python/chat/connectivity-app/requirements.txt
Flask==3.1.0
protobuf==5.29.2
google_auth_oauthlib==1.2.1
google-api-core==2.25.1
google-apps-meet==0.1.16
google-cloud-firestore==2.19.0
